refactor(playground): move sim/ and domain/ layers
Relocate pure-logic modules into layered directories:

sim/
  - sim.ts (WasmSim wrapper, verbatim move)
  - traffic-driver.ts (TrafficDriver, renamed from traffic.ts)

domain/
  - permalink/ (URL codec + seed hash)
  - params/ (tweak drawer logic + RON builder)
  - scenarios/ (embedded RON scenarios + registry)

Each directory has a barrel index.ts. All test imports updated.
Zero logic changes — only file moves and import path rewrites.
