+++ playground/src/domain/permalink/index.ts
@@ -1,0 +1,7 @@
+export {
+  DEFAULT_STATE,
+  decodePermalink,
+  encodePermalink,
+  hashSeedWord,
+  type PermalinkState,
+} from "./permalink";
