+++ playground/src/domain/scenarios/index.ts
@@ -1,0 +1,1 @@
+export { SCENARIOS, scenarioById } from "./scenarios";
