+++ playground/src/sim/index.ts
@@ -1,0 +1,2 @@
+export { Sim, loadWasm } from "./sim";
+export { TrafficDriver, type RiderSpec } from "./traffic-driver";
